% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{sp_color_styles}
\alias{sp_color_styles}
\title{Available coordinated spatial color styles}
\usage{
sp_color_styles()
}
\value{
Table of categorical and continuous style choices.
}
\description{
Available coordinated spatial color styles
}
